--- a/src/nn/nn-vulkan.cpp
+++ b/src/nn/nn-vulkan.cpp
@@ -489,6 +489,10 @@ static std::vector<NnVulkanBatchInfo> buildBatchInfo(NnOpConfig *opConfig, NnVul
     return offset;
 }
 
+static NnUint resolveShaderConstants(const NnOpConfig *opConfig, int* consts) {
+    return 0;
+}
+
 static void resolveShaderGroups(const NnOpConfig *opConfig, const NnUint batchSize, NnUint *groupCount) {
     groupCount[0] = 1;
     groupCount[1] = batchSize;
@@ -599,6 +603,12 @@ NnVulkanDeviceSegment::NnVulkanDeviceSegment(NnVulkanContext *context, NnVulkanD
     std::vector<vk::PipelineShaderStageCreateInfo> shaderCreateInfos(segmentConfig->nOps);
     std::vector<std::vector<NnVulkanBuffer *>> opBuffers(segmentConfig->nOps);
 
+    constexpr NnUint maxConsts = 3;
+    std::vector<NnUint> nConsts(segmentConfig->nOps);
+    std::vector<int> consts(segmentConfig->nOps * maxConsts);
+    std::vector<vk::SpecializationInfo> specInfos(segmentConfig->nOps);
+    std::vector<vk::SpecializationMapEntry> specMapEntries(segmentConfig->nOps * maxConsts);
+    
     for (NnUint opIndex = 0; opIndex < segmentConfig->nOps; opIndex++) {
         NnOpConfig *opConfig = &segmentConfig->ops[opIndex];
         NnSize2D inputSize = data->resolveBufferSize(&opConfig->input);
@@ -610,6 +620,23 @@ NnVulkanDeviceSegment::NnVulkanDeviceSegment(NnVulkanContext *context, NnVulkanD
         );
         const char *shaderFileName = getShaderFileName(opConfig->code, opQuant);
         std::vector<uint32_t> code = readShader(shaderFileName);
+    
+        NnUint nConsts = resolveShaderConstants(opConfig, &consts[opIndex * maxConsts]);
+        if (nConsts > 0) {
+            for (NnUint i = 0; i < nConsts; i++) {
+                specMapEntries[opIndex * maxConsts + i] = vk::SpecializationMapEntry(
+                    i,
+                    sizeof(int) * i,
+                    sizeof(int)
+                );
+            }
+            specInfos[opIndex] = vk::SpecializationInfo(
+                nConsts,
+                &specMapEntries[opIndex * maxConsts],
+                nConsts * sizeof(int),
+                &consts[opIndex * maxConsts]
+            );
+        }
 
         std::vector<NnVulkanBuffer *> &buffers = opBuffers[opIndex];
         buildShaderLayout(buffers, data, segmentData.get(), opIndex, opConfig);
@@ -620,12 +647,14 @@ NnVulkanDeviceSegment::NnVulkanDeviceSegment(NnVulkanContext *context, NnVulkanD
             code.size(),
             code.data()
         );
+
         vk::ShaderModule shaderModule = context->device.createShaderModule(shaderModuleCreateInfo);
         vk::PipelineShaderStageCreateInfo shaderCreateInfo(
             vk::PipelineShaderStageCreateFlags(),
             vk::ShaderStageFlagBits::eCompute,
             shaderModule,
-            "main"
+            "main",
+            nConsts > 0 ? &specInfos[opIndex] : nullptr
         );
 
         shaderModules[opIndex] = shaderModule;