--- a/src/nn/vulkan/matmul-forward-q80-q40-f32.comp
+++ b/src/nn/vulkan/matmul-forward-q80-q40-f32.comp
@@ -5,10 +5,9 @@
 #extension GL_EXT_shader_explicit_arithmetic_types : enable
 
 #define Q80_Q40_BLOCK_SIZE 32
-#define N_THREADS 256
+#define N_THREADS 64
 
-#define N_OUTPUTS_PER_ITER 64
-#define N_THREADS_PER_OUTPUT (N_THREADS / N_OUTPUTS_PER_ITER)
+#define TILE_SIZE 2
 
 layout(local_size_x = N_THREADS, local_size_y = 1, local_size_z = 1) in;
 
@@ -34,13 +33,13 @@ layout(binding = 1) writeonly buffer outputBuffer { float y[]; };
 layout(binding = 2) readonly buffer batchInfosBuffer { BatchInfo infos[]; };
 layout(binding = 3) readonly buffer weightBuffer { BlockQ40 weight[]; };
 
-shared uint sharedStart;
-shared uint sharedEnd;
+shared uint sharedDStart;
+shared uint sharedDEnd;
 shared uint sharedInputOffset;
 shared uint sharedInputSizeX;
 shared uint sharedOutputOffset;
-shared uint sharedInputSizeXPerGroup;
-shared float16_t sums[N_THREADS];
+shared uint sharedInputSizeXPerThread;
+shared float16_t sums[N_THREADS * TILE_SIZE];
 
 void main() {
     const uint threadIndex = gl_LocalInvocationID.x;
@@ -54,58 +53,72 @@ void main() {
         sharedInputOffset = info.inputOffset;
         sharedInputSizeX = info.inputSizeX;
         sharedOutputOffset = info.outputOffset;
-        sharedInputSizeXPerGroup = (sharedInputSizeX + N_THREADS_PER_OUTPUT - 1) / N_THREADS_PER_OUTPUT;
+        sharedInputSizeXPerThread = (sharedInputSizeX + N_THREADS - 1) / N_THREADS;
 
         const uint ySlice = info.outputSizeX / nWorkGroups;
         const uint yRest = info.outputSizeX % nWorkGroups;
-        sharedStart = workGroupIndex * ySlice + (workGroupIndex < yRest ? workGroupIndex : yRest);
-        sharedEnd = sharedStart + ySlice + (workGroupIndex < yRest ? 1 : 0);
+        sharedDStart = workGroupIndex * ySlice + (workGroupIndex < yRest ? workGroupIndex : yRest);
+        sharedDEnd = sharedDStart + ySlice + (workGroupIndex < yRest ? 1 : 0);
     }
 
     barrier();
     memoryBarrierShared();
 
-    const uint dEnd = sharedEnd;
+    const uint dEnd = sharedDEnd;
     const uint inputOffset = sharedInputOffset;
     const uint inputSizeX = sharedInputSizeX;
     const uint outputOffset = sharedOutputOffset;
-    const uint inputSizeXPerGroup = sharedInputSizeXPerGroup;
+    const uint inputSizeXPerThread = sharedInputSizeXPerThread;
 
-    const uint dGroup = threadIndex / N_THREADS_PER_OUTPUT;
-    const uint iGroup = threadIndex % N_THREADS_PER_OUTPUT;
-    const uint iStart = inputSizeXPerGroup * iGroup;
-    const uint iEnd = min(iStart + inputSizeXPerGroup, inputSizeX);
+    const uint iStart = inputSizeXPerThread * threadIndex;
+    const uint iEnd = min(iStart + inputSizeXPerThread, inputSizeX);
 
-    for (uint dBatch = sharedStart; dBatch < dEnd; dBatch += N_OUTPUTS_PER_ITER) {
-        const uint d = dBatch + dGroup;
-        if (d >= dEnd) {
-            break;
+    float16_t xTemp[Q80_Q40_BLOCK_SIZE];
+
+    for (uint d = sharedDStart; d < dEnd; d += TILE_SIZE) {
+        for (uint dt = 0; dt < TILE_SIZE; dt++) {
+            sums[threadIndex * TILE_SIZE + dt] = float16_t(0.0f);
         }
 
-        float16_t sum = float16_t(0.0f);
-        for (uint i = iStart; i < iEnd; i++) {
-            const uint xi = inputOffset + i;
-            const uint wi = d * inputSizeX + i;
-            const float16_t scale = x[xi].d * weight[wi].d;
-            [[unroll]] for (uint j = 0; j < Q80_Q40_BLOCK_SIZE / 2; j++) {
-                sum += (
-                    float16_t(x[xi].qs[j])                          * (float16_t(weight[wi].qs[j] & 0xF) - float16_t(8.0f)) +
-                    float16_t(x[xi].qs[j + Q80_Q40_BLOCK_SIZE / 2]) * (float16_t(weight[wi].qs[j] >>  4) - float16_t(8.0f))
-                ) * scale;
+        for (uint i = iStart; i < iEnd; i += TILE_SIZE) {
+            [[unroll]] for (uint it = 0; it < TILE_SIZE; it++) {
+                const uint xi = inputOffset + i + it;
+                const float16_t xScale = x[xi].d;
+                [[unroll]] for (uint j = 0; j < Q80_Q40_BLOCK_SIZE / 2; j++) {
+                    xTemp[j * 2]     = float16_t(x[xi].qs[j]);
+                    xTemp[j * 2 + 1] = float16_t(x[xi].qs[j + Q80_Q40_BLOCK_SIZE / 2]);
+                }
+
+                [[unroll]] for (uint dt = 0; dt < TILE_SIZE; dt++) {
+                    const uint wi = (d + dt) * inputSizeX + (i + it);
+                    const BlockQ40 wBlock = weight[wi];
+
+                    float16_t s = float16_t(0.0f);
+                    [[unroll]] for (uint j = 0; j < Q80_Q40_BLOCK_SIZE / 2; j++) {
+                        s += (
+                            xTemp[j * 2]     * (float16_t(wBlock.qs[j] & 0xF) - float16_t(8.0f)) +
+                            xTemp[j * 2 + 1] * (float16_t(wBlock.qs[j] >>  4) - float16_t(8.0f))
+                        );
+                    }
+                    sums[threadIndex * TILE_SIZE + dt] += s * xScale * wBlock.d;
+                }
             }
         }
-        sums[threadIndex] = sum;
 
         barrier();
         memoryBarrierShared();
 
-        [[unroll]] for (uint i = N_THREADS_PER_OUTPUT / 2; i > 0; i >>= 1) {
-            if (iGroup < i)
-                sums[threadIndex] += sums[threadIndex + i];
+        [[unroll]] for (uint i = N_THREADS / 2; i > 0; i >>= 1) {
+            for (uint dt = 0; dt < TILE_SIZE; dt++) {
+                if (threadIndex < i) {
+                    sums[threadIndex * TILE_SIZE + dt] += sums[(threadIndex + i) * TILE_SIZE + dt];
+                }
+            }
             barrier();
         }
-        if (iGroup == 0) {
-            y[outputOffset + d] = float(sums[threadIndex]);
+
+        if (threadIndex < TILE_SIZE) {
+            y[outputOffset + d + threadIndex] = float(sums[threadIndex]);
         }
 
         barrier();